requirements.txt
# python -m pip install -r requirements.txt

# Framework
FastAPI==0.128.0

# Server
Uvicorn==0.40.0

# Obsługa HTTP
requests==2.32.5

# Środowisko wirtualne
dotenv==0.9.9

# Modele LLM
openai==2.15.0
google-generativeai==0.8.6
anthropic==0.76.0